# Conflicts:
#	dbms/src/test/java/org/polypheny/db/jdbc/JdbcMetaTest.java
#	plugins/jdbc-adapter-framework/src/main/java/org/polypheny/db/adapter/jdbc/JdbcToEnumerableConverter.java
#	plugins/monetdb-adapter/gradle.properties
#	plugins/postgres-adapter/gradle.properties
#	plugins/proto-interface/src/main/java/org/polypheny/db/protointerface/utils/PolyValueSerializer.java

+++ plugins/sql-language/src/main/java/org/polypheny/db/sql/language/ddl/altertable/SqlAlterTableDropColumn.java
@@ -20,6 +20,7 @@ package org.polypheny.db.sql.language.ddl.altertable;
 import java.util.List;
 import java.util.Objects;
 import org.polypheny.db.catalog.entity.logical.LogicalTable;
+import org.polypheny.db.catalog.exceptions.GenericRuntimeException;
 import org.polypheny.db.catalog.logistic.EntityType;
 import org.polypheny.db.ddl.DdlManager;
 import org.polypheny.db.languages.ParserPos;
@@ -75,14 +76,18 @@ public class SqlAlterTableDropColumn extends SqlAlterTable {
 
     @Override
     public void execute( Context context, Statement statement, QueryParameters parameters ) {
-        LogicalTable catalogTable = getFromCatalog( context, table );
+        LogicalTable catalogTable = getEntityFromCatalog( context, table );
+
+        if ( catalogTable == null ) {
+            throw new GenericRuntimeException( "Entity with name '%s' was not found.", String.join( ".", table.names ) );
+        }
 
         if ( catalogTable.entityType != EntityType.ENTITY && catalogTable.entityType != EntityType.SOURCE ) {
-            throw new RuntimeException( "Not possible to use ALTER TABLE because " + catalogTable.name + " is not a table." );
+            throw new GenericRuntimeException( "Not possible to use ALTER TABLE because %s is not a table.", catalogTable.name );
         }
 
         if ( column.names.size() != 1 ) {
-            throw new RuntimeException( "No FQDN allowed here: " + column.toString() );
+            throw new GenericRuntimeException( "No FQDN allowed here: %s", column );
         }
 
         DdlManager.getInstance().dropColumn( catalogTable, column.getSimple(), statement );
